telegram_bot/requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.1
openai-whisper==20231117
torch>=1.10.0
ffmpeg-python==0.2.0
